refactor: prompt form refactor & minor fixes
