+++ package.json
@@ -37,7 +37,8 @@
     "react-markdown": "^10.1.0",
     "rehype-highlight": "^7.0.2",
     "sonner": "^2.0.6",
-    "tailwind-merge": "^3.3.1"
+    "tailwind-merge": "^3.3.1",
+    "uuid": "^11.1.0"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
