+++ src/app/_components/index.ts
@@ -1,13 +1,11 @@
 export * from "./Controll";
 export * from "./MarkdownPreComponent";
 export * from "./ModelSelector";
-export * from "./PromptForm";
+export * from "./prompt/PromptForm";
 export * from "./RenderMessages";
 export * from "./SetupForm";
 export * from "./ThemeChanger";
 export * from "./Usage";
-export * from "./Spinner";
-export * from "./Spinner";
 export * from "./Spinner";
 
 export * from "./hooks/localStorage.hook";
