+++ src/app/_components/prompt/index.ts
@@ -1,0 +1,3 @@
+export * from "./AttachedImages";
+export * from "./PromptControll";
+export * from "./PromptForm";
